refactor: Separate file for Drive breadcrumb
